+++ packages/react/src/experimental/Navigation/Header/Breadcrumbs/internal/BreadcrumbSkeleton.tsx
@@ -1,0 +1,18 @@
+import { Skeleton } from "@/ui/skeleton"
+import { ComponentPropsWithoutRef, forwardRef } from "react"
+
+/**
+ * Loading skeleton for breadcrumb items
+ */
+const BreadcrumbSkeleton = forwardRef<
+  HTMLDivElement,
+  ComponentPropsWithoutRef<"div">
+>((props, ref) => (
+  <div ref={ref} className="px-1.5" {...props}>
+    <Skeleton className="h-4 w-24" aria-hidden="true" />
+  </div>
+))
+
+BreadcrumbSkeleton.displayName = "BreadcrumbSkeleton"
+
+export { BreadcrumbSkeleton }
